[架构更新] 云函数迁移 1. 增加 迁移服务程序登录、检票、验票、票券列表、购票、退票操作到云函数 2. 新增 一些加载提示动画和提示消息 3. 新增 删除票券的确认 4. 修改 修改小程序API接口为云函数 5. 修改 模块独立--验票和检票函数移动到对应模块，与主界面解耦
 - TODO 增加 定时出发云函数处理票券过期问题
 - TODO 增加 小程序页面判断日期来标记过期
