+++ .idea/dictionaries/sam.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="sam">
+    <words>
+      <w>muumlover</w>
+    </words>
+  </dictionary>
+</component>
