+++ client/src/common/wxCloud.js
@@ -1,0 +1,11 @@
+import Taro from "@tarojs/taro";
+
+export const CloudCall = (name, data) => new Promise((resolve, reject) => {
+  Taro.cloud.callFunction({name, data}).then(res => {
+    console.log(res.result);
+    resolve(res.result);
+  }).catch(e => {
+    console.log(e);
+    reject(e)
+  });
+});
