+++ client/src/component/modal_ticket_checked/index.scss
