+++ client/src/component/modal_ticket_purchase/index.scss
