+++ client/src/component/ticket_list/index.scss
