+++ client/src/component/ticket_scan/index.scss
