+++ client/src/component/user_info/index.scss
