+++ client/src/pages/qrcode_show/index.scss
