+++ client/src/pages/ticket_show/index.scss
